+++ printing/lib/printing.dart
@@ -17,6 +17,7 @@
 library printing;
 
 import 'dart:async';
+import 'dart:io';
 import 'dart:typed_data';
 import 'dart:ui' as ui;